pyproject.toml
[tool.poetry]
name = "imagesmacker"
version = "0.0.0"
description = ""
authors = ["whinee"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
rich = "^13.3.5"
msgpack = "^1.0.5"
openpyxl = "^3.1.2"
pyyaml = "^6.0"
regex = "^2023.8.8"
pillow = "^10.0.1"
qrcode = {extras = ["pil"], version = "^7.4.2"}
alltheutils = {git = "https://github.com/klikasnap/AllTheUtils.git"}
pydantic = "^2.8.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.1.1"
no-implicit-optional = "^1.3"
ruff = "^0.0.251"

[tool.mypy]
disable_error_code = [
    "import",
    "annotation-unchecked",
    "attr-defined",
    "no-untyped-call",
    "no-untyped-def",
    "type-arg",
]
strict = true
exclude = "dev"

[tool.ruff]
select = [
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "D",
    # "DTZ",
    "E",
    "F",
    "I",
    "INP",
    "N",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "UP"
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "ANN",
    "B",
    "C4",
    "COM",
    "D",
    "E",
    "F",
    "I",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "UP"
]
unfixable = []

ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN201",
    "ANN202",
    "ANN401",
    "B008",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D202",
    "D203",
    "D212",
    "D401",
    "D404",
    "D406",
    "D407",
    "D417",
    "E501",
    "F722",
    "I002",
    "N812",
    "N818",
    "Q000",
    "RET503",
    "S101",
    "UP007",
    "UP032",
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "pyenv",
    "venv",
]

line-length = 88

target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 5

[tool.ruff.pycodestyle]
ignore-overlong-task-comments = false